Starting work on codec.go
